+++ src/App.tsx
@@ -1,22 +1,32 @@
 import { ThemeProvider } from './libs/context/themeContext';
-import Mission from './scenes/Mission';
-import {
-  BrowserRouter as Router,
-  Route,
-  Routes,
-  Link
-} from "react-router-dom";
+import { MissionWizardProvider } from './libs/context/MissionWizardContext';
+import MissionCard from './libs/components/MissionCard';
+import missions from "./helpers/missions.json";
+import "./libs/scss/index.scss";
 
 function App() {
-
   return (
     <ThemeProvider>
-      <Router>
-        <Link to="/mission">Mission</Link>
-        <Routes>
-          <Route path="/mission" element={<Mission/>} />
-        </Routes>
-      </Router>
+      <MissionWizardProvider>
+        <header className='list-missions__header'>
+          <h1>Trouve ta mission</h1>
+        </header>
+        <main className="missions_slider">
+          {missions.map(missionsData => {
+            return <MissionCard missionData= {missionsData}/>;
+          })}
+        </main>
+        <footer>
+          <ul>
+            <li>
+              <p>Missions</p>
+            </li>
+            <li>
+              <p>Deconnexion</p>
+            </li>
+          </ul>
+        </footer>
+      </MissionWizardProvider>
     </ThemeProvider>
 
   );