+++ src/libs/context/MissionWizardContext.tsx
@@ -1,15 +1,18 @@
 import React, { ReactNode, useState } from "react";
-import { IsMissionInterface } from "../Interface";
+import { IsMissionInterface } from "./Interface";
 
 export const MissionWizardContext = React.createContext<IsMissionInterface | null>(null);
 
 export const MissionWizardProvider = ({ children }: {children: ReactNode}) => {
-     const defaultValues = {
-        title: "Test",
-        steps: {
-            step1: "test"
-        }
-     }
+    const [title, setTitle] = useState<string | null>(null);
+    const [steps, setSteps] = useState<Array<object | null> >([]);
+    
+    const defaultValues = {
+        title,
+        setTitle, 
+        steps,
+        setSteps
+    }
 
      return (
         <MissionWizardContext.Provider value={defaultValues}>